completed : order & inventory system

+++ 02-OrderAndInventoryManagement/OrderAndInventoryManagement/src/main/java/com/priyanshu/orderAndInventoryManagement/OrderAndInventoryManagement/dto/inventory/InventoryPageResponse.java
@@ -4,7 +4,7 @@ public record InventoryPageResponse(
         Long productId,
         String name,
         String description,
-        Double price,
+        BigDecimal price,
         Integer quantity
 
 ) {
