+++ 02-OrderAndInventoryManagement/OrderAndInventoryManagement/src/main/java/com/priyanshu/orderAndInventoryManagement/OrderAndInventoryManagement/dto/products/ProductResponse.java
@@ -1,9 +1,11 @@
 package com.priyanshu.orderAndInventoryManagement.OrderAndInventoryManagement.dto.products;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
         Long id,
         String name,
-        double price,
+        BigDecimal price,
         String description
 
 ) {
