+++ 02-OrderAndInventoryManagement/OrderAndInventoryManagement/src/main/java/com/priyanshu/orderAndInventoryManagement/OrderAndInventoryManagement/entity/Order.java
@@ -7,6 +7,7 @@ import lombok.experimental.FieldDefaults;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Entity
@@ -31,7 +32,7 @@ public class Order {
     String shippingAddress;
 
     @Column(nullable = false)
-    Double totalAmount;
+    BigDecimal totalAmount;
 
     @CreationTimestamp
     Instant created_at;
